\name{gFunctions}
\alias{gFunctions}
\title{Load a Set of R Function Files}

\description{Loads or reloads the files listed within the program to simplify loading the GCAF pseudo-package.}

\usage{source("~/Documents/My Dropbox/ISB/Notebook/GCAF/R/gFunctions.r")}

\details{User should edit when new functions are added to the pseudo-package.}

\author{Goodwin Gibbins}
